requirements.txt
appdirs==1.4.4
colorama==0.4.4
distlib==0.3.1
filelock==3.0.12
Pillow==8.1.2
PyQRCode==1.2.1
PySide2==5.15.2
pywin32==300
qrcode==6.1
shiboken2==5.15.2
six==1.15.0
virtualenv==20.4.2
zebra==0.1.0
zpl==0.1.8
